Consolidated createInput, createButton and createLabel functions into single file.
New location client/src/js/utilities.js.
